requirements.txt
Babel==1.3
Flask==0.10.1
Flask-Admin==1.0.7
Flask-BabelEx==0.9.1
Flask-Login==0.2.9
Flask-Migrate==1.2.0
Flask-SQLAlchemy==1.0
Flask-Script==0.6.7
Flask-WTF==0.9.4
GitPython==0.3.2.RC1
Jinja==1.2
Jinja2==2.7.2
Mako==0.9.1
MarkupSafe==0.18
MySQL-python==1.2.5
Pillow==2.3.0
SQLAlchemy==0.9.3
WTForms==1.0.5
Werkzeug==0.9.4
alembic==0.6.3
argparse==1.2.1
async==0.6.1
gitdb==0.5.4
itsdangerous==0.23
pytz==2013.9
smmap==0.8.2
speaklater==1.3
wsgiref==0.1.2
